Search components project structure updated.
